estimateDict -> ppeDict
